requirements.txt
pytest==8.4.2
pytest-cov==6.3.0
pytest-mock==3.15.0
pylint==3.3.8
black==25.1.0
mypy==1.17.1
mypy-extensions==1.1.0
urllib3==2.6.0
cryptography==45.0.7
jsonschema==4.25.1
PyJWT==2.10.1
requests==2.32.5
boto3==1.40.25
confluent-kafka==2.12.1
# psycopg2-binary==2.9.10 # Ideal for local development, but not for long-term production use
psycopg2==2.9.10
botocore==1.42.4
